Implement new vertical timeline.
