+++ package.json
@@ -17,11 +17,12 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "react-sound": "^1.2.0",
     "react-spinners": "^0.11.0",
     "react-spring": "^9.0.0",
     "react-typist": "^2.0.5",
+    "react-xarrows": "^2.0.2",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
