+++ src/components/scroll-anim-container/ScrollAnimContainer.jsx
@@ -15,7 +15,12 @@ import ScrollAnimSection from './ScrollAnimSection';
 
 // const SHOW_BUTTON_DELAY_MS = 800;
 
-const ScrollAnimContainer = ({ children, horizontal, showChild, onChange }) => {
+const ScrollAnimContainer = ({
+  children,
+  horizontal,
+  showChildIdx,
+  onChange,
+}) => {
   const [childrenList, setChildrenList] = useState(
     Array.isArray(children)
       ? children.map((elem, idx) => ({
@@ -46,7 +51,6 @@ const ScrollAnimContainer = ({ children, horizontal, showChild, onChange }) => {
   const [currChildIdx, setCurrChildIdx] = useState(0);
   const [renderNextButton, setRenderNextButton] = useState(false);
   const [renderPrevButton, setRenderPrevButton] = useState(false);
-  const [nameToIdxMap, setNameToIdxMap] = useState(new Map());
   const [showLabels, setShowLabels] = useState(true);
 
   // animation for next button
@@ -104,33 +108,28 @@ const ScrollAnimContainer = ({ children, horizontal, showChild, onChange }) => {
     });
   };
 
-  const setCurrent = (idx, name) => {
-    onChange(name);
+  const setCurrent = (idx) => {
+    onChange(idx);
     setPositionOfChildAt(idx, 0);
     setCurrChildIdx(idx);
   };
 
-  const moveToChild = (name) => {
-    const showChildIdx = nameToIdxMap.get(name);
-    if (currChildIdx > showChildIdx) {
-      for (let i = showChildIdx + 1; i <= currChildIdx; i += 1) {
+  const moveToChildAtIdx = (idx) => {
+    if (currChildIdx > idx) {
+      for (let i = idx + 1; i <= currChildIdx; i += 1) {
         setPositionOfChildAt(i, 1);
       }
-    } else if (currChildIdx < showChildIdx) {
-      for (let i = showChildIdx - 1; i >= currChildIdx; i -= 1) {
+    } else if (currChildIdx < idx) {
+      for (let i = idx - 1; i >= currChildIdx; i -= 1) {
         setPositionOfChildAt(i, -1);
       }
     }
-    setCurrent(showChildIdx, name);
+    setCurrent(idx);
   };
 
   useEffect(() => {
-    if (showChild && nameToIdxMap.has(showChild)) {
-      if (currChildIdx !== nameToIdxMap.get(showChild)) {
-        moveToChild(showChild);
-      }
-    }
-  }, [showChild]);
+    moveToChildAtIdx(showChildIdx);
+  }, [showChildIdx]);
 
   const setNextButtonConfigOfChildAt = (childIdx, cfg) => {
     if (cfg.show && childIdx === currChildIdx) {
@@ -150,15 +149,6 @@ const ScrollAnimContainer = ({ children, horizontal, showChild, onChange }) => {
     setChildrenList((list) => {
       const newList = list;
       newList[childIdx].prevButtonConfig = cfg;
-      return [...newList];
-    });
-  };
-
-  const mapNameOfChildAt = (childIdx, name) => {
-    setNameToIdxMap((prev) => new Map(prev).set(name, childIdx));
-    setChildrenList((prev) => {
-      const newList = prev;
-      newList[childIdx].name = name;
       return [...newList];
     });
   };
@@ -196,7 +186,6 @@ const ScrollAnimContainer = ({ children, horizontal, showChild, onChange }) => {
             horizontal={horizontal}
             first={idx === 0}
             last={idx === childrenList.length - 1}
-            setName={(name) => mapNameOfChildAt(idx, name)}
             setNextButtonConfig={(cfg) =>
               setNextButtonConfigOfChildAt(idx, cfg)
             }
@@ -301,14 +290,14 @@ ScrollAnimContainer.propTypes = {
     PropTypes.element,
   ]),
   horizontal: PropTypes.bool, // set to true to change orientation to horizontal
-  showChild: PropTypes.string, // name of child you want to show
+  showChildIdx: PropTypes.number, // name of child you want to show
   onChange: PropTypes.func, // callback function on state change
 };
 
 ScrollAnimContainer.defaultProps = {
   children: [],
   horizontal: false,
-  showChild: '',
+  showChildIdx: 0,
   onChange: () => {},
 };
