+++ src/components/scroll-anim-container/ScrollAnimItem.jsx
@@ -14,8 +14,6 @@ const ScrollAnimItem = ({
   shown,
   setNextButtonConfig,
   setPrevButtonConfig,
-  setName,
-  name,
 }) => {
   const [nextScrollEnabled, setNextScrollEnabled] = useState(
     nextButton.scrollable
@@ -34,12 +32,6 @@ const ScrollAnimItem = ({
   useEffect(() => {
     setPrevButtonConfig(prevConfig);
   }, [prevConfig]);
-
-  useEffect(() => {
-    if (name) {
-      setName(name);
-    }
-  }, []);
 
   // checks whether current scroll is at the bottom of the given element
   const isAtElementBottom = (element) => {
@@ -80,7 +72,6 @@ const ScrollAnimItem = ({
 
 ScrollAnimItem.propTypes = {
   content: PropTypes.element,
-  name: PropTypes.string, // name of this item
   nextButton: PropTypes.shape({
     show: PropTypes.bool, // set to false to hide this button
     label: PropTypes.string, // put a label next to the button
@@ -99,7 +90,6 @@ ScrollAnimItem.propTypes = {
   shown: PropTypes.bool, // true if this current element is shown
   setNextButtonConfig: PropTypes.func,
   setPrevButtonConfig: PropTypes.func,
-  setName: PropTypes.func,
 };
 
 ScrollAnimItem.defaultProps = {
@@ -109,7 +99,6 @@ ScrollAnimItem.defaultProps = {
       <p>This is a section</p>{' '}
     </div>
   ),
-  name: '',
   nextButton: {
     show: true,
     label: '',
@@ -128,7 +117,6 @@ ScrollAnimItem.defaultProps = {
   shown: false,
   setNextButtonConfig: () => {},
   setPrevButtonConfig: () => {},
-  setName: () => {},
 };
 
 export default ScrollAnimItem;
