+++ src/components/timeline/EventDetails.css
@@ -12,7 +12,7 @@
 }
 
 .event-details-content {
-  width: 85%;
+  width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
@@ -25,8 +25,14 @@
   color: black;
   font-size: 50px;
   text-align: left;
-  margin-right: 10px;
-  padding-left: 20px;
+  margin-right: 20px;
+  margin-left: 20px;
+  width: 100%;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+}
+
+.event-details-date-divider {
+  margin-top: 0;
 }
 
 .event-details-title {
@@ -40,6 +46,7 @@
   flex-direction: row-reverse;
   align-items: center;
   justify-content: flex-end;
+  padding-top: 14px;
 }
 
 .event-details-svg {
@@ -55,8 +62,8 @@
   padding-left: 20px;
   padding-right: 20px;
   padding-bottom: 20px;
+  padding-top: 10px;
   text-align: left;
-  margin-top: -10px;
 }
 
 .event-details-learn-more {
@@ -79,7 +86,7 @@
 
 .line2 {
   height: 20px;
-  width: 60%;
+  width: 100%;
   background-color: black;
   margin-left: -10px;
 }
